Update get_printable_seed to give padded base64
Move some comments around, and allow for one extra character in
the seed option (one more than MAX_B64_RNG_SEED_LEN) but *only*
if the last character is '='. Otherwise, the decoded byte string
would exceed 32 bytes.

An unpadded string in the rcfile will still give the same RNG
seed as if the missing '=' chars were there, but it's nice to use
correct base64. Currently MAX_B64_RNG_SEED_LEN is one byte fewer
than the actual max, but to obtain 32 bytes on decode, the b64
string must be MAX_B64_RNG_SEED_LEN + 1 with '=' as the final byte.

Because the RNG seed itself is always 32 bytes, if the user-set b64
seed decodes to fewer than 32 bytes, the remaining bytes are padded
with zeroes. This means seed=foobar== should give exactly the same
result as seed=foobarAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=
This is not really a problem as far as I'm concerned, however...

There is still an issue in b64_decode(): if a padding char or end
of string is found, the lower n bits of the final sextet become the
upper n bits of an additional octet. If the input is valid base64
this will result in double null-termination, but the input is not
checked to ensure terminal chunks are actually correct base64.
This means certain strings allowed by the current validation will
overflow the 32-byte RNG buffer with something other than null.

TODO: handle base64 string-termination correctly and add proper
input validation to ensure these bugs don't happen.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>