Fix overflow cause by adding leftover sextet bits
Prevent unwanted extra bytes being added by the decoder, the
offending bit of code was an if (dest_bitdex) {
dest[dest_index++] = u8_val; } that came after the main loop.
However, if the input is valid base64, these bits should be 0
and followed by padding '=' chars. Remove.

Also remove an increment which resulted in the returned length
including the null-termination character. Now return is the
string length excluding the null-termination character.

Add more rigorous input validation to the base64 code.

Unfortunately, this means confusion for users, beacuse now
the allowed chars are a-zA-Z0-9+/ BUT they have to know which
terminating sequences are valid base64, which is not user-friendly
at all!

I think the easiest thing would be just to pad the user
input base64 seeds with some combination of A and = as padding.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>