Add some internal base64 hax to pad user-input
To deal with trailing bits in user supplied seeds in a non-lossy
manner, some internal extra padding is added to certain cases.
e.g. "foobar" is invalid base64, even with padding = chars added.
"foobaQ==" is valid base64, "foobar==" is not. So, internally,
"foobar" would be padded to become "foobarA=", which is valid
and doesn't lose the low bits of the sextet corresponding to "r".

However, we don't want the user getting different things printed
out in logs or shown in options than what they configured.
So, add a global to save the user-supplied seed from the config and
have get_printable_seed() simply return that if flags.setseed is
set.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>