add rough prototype for seedlist functionality
This should be squashed later, as feature does not work correctly.
Compiles at least. Anyway, we have two problems as of this commit.
1) g.plname is an empty string when the seedlist option is processed.
2) seedlist option is processed both on game init, and save load.

Problem 1) means that all players will share the same seedlist-index
file named "-seedlist" instead of "playername-seedlist".
Problem 2) means that save/reload will increment the counter and
progress player (or players, with problem 1 still in effect) through the
seed list too quickly and when the seeds run out, they cannot play.

I dunno how to solve this yet. Problem 2) may be dealt with by ensuring
that seedlist is a birth-only option, which seed should have been as
well anyway. I'll want to check how the seed option is dealt with, and
how set-seed deals with save/reload in general *anyway*.
As for problem 1), I want to find out how dgamelaunch passes the player
name to NetHack in server env., and how/why g.plname is still not set
if I have name:foo set in my .nethackrc *prior* to seedlist:rednaga.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>